Выносит отрисовку спиннера и ошибки загрузки в отдельный route
